pyproject.toml
[tool.poetry]
name = "electricity-consumption-forecasting-mlops"
version = "0.1.0"
description = ""
authors = ["Dmitrii Chebotarev <chebotarev.da@phystech.edu>"]
readme = "README.md"
packages = [{include = "electricity_consumption_forecasting_mlops"}]

[tool.poetry.dependencies]
pandas = "2.*"
python = "^3.10"
torch = "1.13.*"
torchvision = "*"
numpy = "^1.26.4"
tqdm = "^4.66.2"
torchmetrics = "^1.3.1"
dvc = "^3.50.0"
dvc-gdrive = "^3.0.1"
hydra-core = "^1.3.2"
scikit-learn = "^1.4.2"
lightning = "^2.2.2"
mlflow = "^2.12.1"
holidays = "^0.47"
scipy = "^1.13.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"

[tool.black]
line-length = 90
target-version = ["py310"]

[tool.isort]
src_paths = ["ds_project"]
profile = "black"
line_length = 90
lines_after_imports = 2

[tool.flake8]
max-line-length = 90
ignore = ['E203', 'E501', 'W503', 'B905', 'B950']
max-complexity = 12
select = ['B', 'C', 'E', 'F', 'W', 'B9']
per-file-ignores = [
    # for easier imports to __init__ without __all__
    '**/__init__.py: F401',
    # file to define custom types
    'types.py: F401',
]
count = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
